Extract getSafeInfo
